average changed to a double using static cast

+++ main.cpp
@@ -39,7 +39,7 @@ int main() {
       //display to user 
       cout << "The sum of the valid data is: " << sumData << endl;
       cout << "The number of valid data points is: " << counter << endl;
-      cout << "The average of the valid data is: " << ((sumData) / counter) << endl;
+      cout << "The average of the valid data is: " << (static_cast<double>(sumData) / counter) << endl;
   
   return 0;
